staticfiles/staticfiles/requirements.txt
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-easy-pdf==0.1.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djoser==2.1.0
formatting==0.0.1
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
gunicorn==19.9.0
html5lib==1.1
httpie==2.6.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.1
MarkupSafe==2.0.1
oauthlib==3.1.1
pdfkit==1.0.0
permission==0.4.1
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
PyPDF2==1.26.0
PySocks==1.7.1
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.3
redis==4.0.0
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
wkhtmltopdf==0.2
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
